Skip hasSupport check for List type
Currently, when user applies a List type resource, there is an unnecessary
request to customresourcedefinitions endpoint by validation. The reason of this
request is that validation can not find List type in OpenAPI schema and
behaves this type as CRD.

This PR returns unsupported when List type is applied.

In terms of functionality, there is no bug or any issue, because
eventually `schemaValidation` detects this List and executes `validateList`
function. Only issue is unnecessary CRD enpoint request.

Kubernetes-commit: e78d05daec30a757b3817bbdc78f5ea0c1c7ea17